Move info components to specific files inside "info" directory
