+++ src/core/components/info/index.js
@@ -1,0 +1,5 @@
+export { default as InfoBasePath } from "./info-base-path"
+export { default as Contact } from "./contact"
+export { default as License } from "./license"
+export { default as InfoUrl } from "./info-url"
+export { default as Info } from "./info"
