+++ src/core/components/info/info.jsx
@@ -2,95 +2,9 @@ import React from "react"
 import PropTypes from "prop-types"
 import { fromJS } from "immutable"
 import ImPropTypes from "react-immutable-proptypes"
+
 import { sanitizeUrl } from "core/utils"
-
-
-export class InfoBasePath extends React.Component {
-  static propTypes = {
-    host: PropTypes.string,
-    basePath: PropTypes.string
-  }
-
-  render() {
-    let { host, basePath } = this.props
-
-    return (
-      <pre className="base-url">
-        [ Base URL: {host}{basePath} ]
-      </pre>
-    )
-  }
-}
-
-
-class Contact extends React.Component {
-  static propTypes = {
-    data: PropTypes.object,
-    getComponent: PropTypes.func.isRequired
-  }
-
-  render(){
-    let { data, getComponent } = this.props
-    let name = data.get("name") || "the developer"
-    let url = data.get("url")
-    let email = data.get("email")
-
-    const Link = getComponent("Link")
-
-    return (
-      <div className="info__contact">
-        { url && <div><Link href={ sanitizeUrl(url) } target="_blank">{ name } - Website</Link></div> }
-        { email &&
-          <Link href={sanitizeUrl(`mailto:${email}`)}>
-            { url ? `Send email to ${name}` : `Contact ${name}`}
-          </Link>
-        }
-      </div>
-    )
-  }
-}
-
-class License extends React.Component {
-  static propTypes = {
-    license: PropTypes.object,
-    getComponent: PropTypes.func.isRequired
-
-  }
-
-  render(){
-    let { license, getComponent } = this.props
-
-    const Link = getComponent("Link")
-  
-    let name = license.get("name") || "License"
-    let url = license.get("url")
-
-    return (
-      <div className="info__license">
-        {
-          url ? <Link target="_blank" href={ sanitizeUrl(url) }>{ name }</Link>
-        : <span>{ name }</span>
-        }
-      </div>
-    )
-  }
-}
-
-export class InfoUrl extends React.PureComponent {
-  static propTypes = {
-    url: PropTypes.string.isRequired,
-    getComponent: PropTypes.func.isRequired
-  }
-
-  
-  render() {
-    const { url, getComponent } = this.props
-
-    const Link = getComponent("Link")
-
-    return <Link target="_blank" href={ sanitizeUrl(url) }><span className="url"> { url } </span></Link>
-  }
-}
+import { Contact, License } from "components/info"
 
 export default class Info extends React.Component {
   static propTypes = {
@@ -149,11 +63,4 @@ export default class Info extends React.Component {
     )
   }
 
-}
-
-Info.propTypes = {
-  title: PropTypes.any,
-  description: PropTypes.any,
-  version: PropTypes.any,
-  url: PropTypes.string
 }
